fix issues related to floating number, fix remove products, enhance remove category

+++ src/category/service/category.service.ts
@@ -1,12 +1,15 @@
-import { Injectable, NotFoundException } from "@nestjs/common";
+import { BadRequestException, Injectable, NotFoundException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Category } from "../model/category.entity";
 import { Repository } from "typeorm";
+import { Product } from "src/product/model/product.entity";
 @Injectable()
 export class CategoryService {
   constructor(
     @InjectRepository(Category)
     private categoryRepository: Repository<Category>,
+    @InjectRepository(Product)
+    private productRepository: Repository<Product>,
   ) {}
 
 
@@ -32,6 +35,19 @@ export class CategoryService {
   }
 
   async remove(id: number): Promise<boolean> {
+
+    const products = await this.productRepository.count({
+      where : {
+          category_id : id
+      }
+    })
+
+    if (products > 0) {
+      throw new BadRequestException([
+        'You must delete products assigned to this category',
+      ]);
+    }
+
     return (await this.categoryRepository.delete(id)).affected > 0;
   }
